Move Gunicorn conf tests to separate module
br3ndonland/inboard#11

These tests weren't really testing the start script, so they shouldn't
be in test_start.py. Moving the tests to test_gunicorn_conf.py also
helps emulate the directory structure of the source code, so tests are
more intuitively found.
